# docs/integration-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Умные заметки",
    "description": "Структурирование мыслей. По умолчанию работает имитация ИИ без сетевых запросов.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/internal/v1/telegram/link-request": {
      "post": {
        "tags": [
          "internal-v1"
        ],
        "summary": "Link Request",
        "operationId": "link_request_internal_v1_telegram_link_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramLink"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                },
                "examples": {
                  "empty_input": {
                    "value": {
                      "error": {
                        "code": "empty_input",
                        "message": "В записи нет текста"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000109"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                },
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "Недопустимый сервисный секрет"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000112"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                },
                "examples": {
                  "telegram_not_linked": {
                    "value": {
                      "error": {
                        "code": "telegram_not_linked",
                        "message": "Сначала подтвердите связь в веб-приложении"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000104"
                    }
                  },
                  "action_forbidden": {
                    "value": {
                      "error": {
                        "code": "action_forbidden",
                        "message": "Кнопка принадлежит другому аккаунту"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000110"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                },
                "examples": {
                  "link_expired": {
                    "value": {
                      "error": {
                        "code": "link_expired",
                        "message": "Ссылка истекла"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000105"
                    }
                  },
                  "link_conflict": {
                    "value": {
                      "error": {
                        "code": "link_conflict",
                        "message": "Telegram или аккаунт уже связан"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000106"
                    }
                  },
                  "invalid_link": {
                    "value": {
                      "error": {
                        "code": "invalid_link",
                        "message": "Ссылка недействительна"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000107"
                    }
                  },
                  "action_expired": {
                    "value": {
                      "error": {
                        "code": "action_expired",
                        "message": "Кнопка недействительна или устарела"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000111"
                    }
                  },
                  "conflict": {
                    "value": {
                      "error": {
                        "code": "conflict",
                        "message": "Update уже использован с другим содержимым"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000114"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                },
                "examples": {
                  "input_too_large": {
                    "value": {
                      "error": {
                        "code": "input_too_large",
                        "message": "Запись слишком большая"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000108"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                },
                "examples": {
                  "invalid_input": {
                    "value": {
                      "error": {
                        "code": "invalid_input",
                        "message": "Проверьте формат и длину ввода"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000113"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                },
                "examples": {
                  "rate_limited": {
                    "value": {
                      "error": {
                        "code": "rate_limited",
                        "message": "Слишком много записей ожидают обработки"
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000115"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationError"
                },
                "examples": {
                  "unavailable": {
                    "value": {
                      "error": {
                        "code": "unavailable",
                        "message": "Временная ошибка. Повторите ту же операцию."
                      },
                      "operation_id": "00000000-0000-4000-8000-000000000116"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "InternalServiceToken": []
          }
        ],
        "x-implementation-status": "implemented"
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-implementation-status": "implemented"
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-implementation-status": "implemented"
      }
    },
...
    "/api/v1/captures/text": {
      "post": {
        "summary": "Create Capture",
        "operationId": "create_capture_api_v1_captures_text_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextCapture"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-implementation-status": "implemented"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IntegrationError": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/IntegrationErrorDetail"
          },
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error",
          "operation_id"
        ],
        "title": "IntegrationError"
      },
...
      "LinkResponse": {
        "properties": {
          "link_request_id": {
            "type": "string",
            "title": "Link Request Id"
          },
          "status": {
            "type": "string",
            "const": "pending",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "link_request_id",
          "status"
        ],
        "title": "LinkResponse"
      },
...
      "TelegramLink": {
        "properties": {
          "bot_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Bot Id"
          },
          "update_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Update Id"
          },
          "telegram_user_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Telegram User Id"
          },
          "chat_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Chat Id"
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 22,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bot_id",
          "update_id",
          "telegram_user_id",
          "chat_id",
          "code"
        ],
        "title": "TelegramLink"
      },
...
    },
    "securitySchemes": {
      "InternalServiceToken": {
        "type": "http",
        "scheme": "bearer"
      },
      "BrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "notes_session"
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "beresta_admin"
      }
    }
  }
}